+++ src/MemStat.Net/IMemoryUsageFinder.cs
@@ -1,0 +1,9 @@
+using ResultBoxes;
+namespace MemStat.Net;
+
+public interface IMemoryUsageFinder
+{
+    public ResultBox<UnitValue> ReceiveCurrentMemoryUsage();
+    public ResultBox<double> GetTotalMemoryUsage();
+    public ResultBox<double> GetMemoryUsagePercentage();
+}
