+++ src/Sekiban.Core/Usecase/SekibanUsecase.cs
